feat: option to only allow new records on resource
